* call `onChange` when find-replace is called in order to trigger autosave, enable save button and other hooks * Extract common functions into a separate file to avoid circular dependencies
